feat(cli): add progress module and CLI tests for unified CLI module
Move progress display to scylla.cli.progress as the canonical location,
with scylla.e2e.progress retained as a backward-compatibility shim.
Add test_cli.py for CLI group/command tests and test_progress.py for
the relocated progress module.

Closes #1683
